chore: update chart axis
