+++ src/pages/StatisticsChart/index.tsx
@@ -9,6 +9,7 @@ import { StateWithDispatch } from '../../contexts/providers/reducer'
 import { AppContext } from '../../contexts/providers'
 import i18n from '../../utils/i18n'
 import Loading from '../../components/Loading'
+import { parseInterval, handleAxis } from '../../utils/chart'
 
 const ChartPanel = styled.div`
   margin: 0 10% 30px 10%;
@@ -54,44 +55,6 @@ const LoadingPanel = styled.div`
   }
 `
 
-const handleAxis = (value: BigNumber) => {
-  if (value.isNaN() || value.isZero()) return '0'
-  const kv = value.dividedBy(1000)
-  const mv = kv.dividedBy(1000)
-  const gv = mv.dividedBy(1000)
-  const tv = gv.dividedBy(1000)
-  const pv = tv.dividedBy(1000)
-  const ev = pv.dividedBy(1000)
-  const zv = ev.dividedBy(1000)
-  const yv = zv.dividedBy(1000)
-
-  if (yv.isGreaterThanOrEqualTo(1)) {
-    return `${yv.toFixed()}Y`
-  }
-  if (zv.isGreaterThanOrEqualTo(1)) {
-    return `${zv.toFixed()}Z`
-  }
-  if (ev.isGreaterThanOrEqualTo(1)) {
-    return `${ev.toFixed()}E`
-  }
-  if (pv.isGreaterThanOrEqualTo(1)) {
-    return `${pv.toFixed()}P`
-  }
-  if (tv.isGreaterThanOrEqualTo(1)) {
-    return `${tv.toFixed()}T`
-  }
-  if (gv.isGreaterThanOrEqualTo(1)) {
-    return `${gv.toFixed()}G`
-  }
-  if (mv.isGreaterThanOrEqualTo(1)) {
-    return `${mv.toFixed()}M`
-  }
-  if (kv.isGreaterThanOrEqualTo(1)) {
-    return `${kv.toFixed()}K`
-  }
-  return `${value.toFixed()}`
-}
-
 const uncleRateScale = () => {
   return {
     uncleRate: {
@@ -103,6 +66,13 @@ const uncleRateScale = () => {
 
 export default ({ dispatch }: React.PropsWithoutRef<StateWithDispatch>) => {
   const { statisticsChartData, statisticsUncleRates } = useContext(AppContext)
+
+  let min = 0
+  let max = 0
+  if (statisticsChartData && statisticsChartData.length > 0) {
+    min = statisticsChartData[0].blockNumber
+    max = statisticsChartData[statisticsChartData.length - 1].blockNumber
+  }
 
   const scale = () => {
     return {
@@ -119,10 +89,10 @@ export default ({ dispatch }: React.PropsWithoutRef<StateWithDispatch>) => {
         alias: i18n.t('block.epoch_number'),
       },
       blockNumber: {
-        min: statisticsChartData[0].blockNumber,
-        max: statisticsChartData[statisticsChartData.length - 1].blockNumber,
+        min,
+        max,
         alias: i18n.t('block.block_number'),
-        tickCount: isMobile() ? 6 : 20,
+        tickInterval: parseInterval(max, min),
       },
     }
   }
