+++ src/utils/chart.ts
@@ -1,0 +1,48 @@
+import BigNumber from 'bignumber.js'
+import { isMobile } from './screen'
+
+export const handleAxis = (value: BigNumber) => {
+  if (value.isNaN() || value.isZero()) return '0'
+  const kv = value.dividedBy(1000)
+  const mv = kv.dividedBy(1000)
+  const gv = mv.dividedBy(1000)
+  const tv = gv.dividedBy(1000)
+  const pv = tv.dividedBy(1000)
+  const ev = pv.dividedBy(1000)
+  const zv = ev.dividedBy(1000)
+  const yv = zv.dividedBy(1000)
+
+  if (yv.isGreaterThanOrEqualTo(1)) {
+    return `${yv.toFixed()}Y`
+  }
+  if (zv.isGreaterThanOrEqualTo(1)) {
+    return `${zv.toFixed()}Z`
+  }
+  if (ev.isGreaterThanOrEqualTo(1)) {
+    return `${ev.toFixed()}E`
+  }
+  if (pv.isGreaterThanOrEqualTo(1)) {
+    return `${pv.toFixed()}P`
+  }
+  if (tv.isGreaterThanOrEqualTo(1)) {
+    return `${tv.toFixed()}T`
+  }
+  if (gv.isGreaterThanOrEqualTo(1)) {
+    return `${gv.toFixed()}G`
+  }
+  if (mv.isGreaterThanOrEqualTo(1)) {
+    return `${mv.toFixed()}M`
+  }
+  if (kv.isGreaterThanOrEqualTo(1)) {
+    return `${kv.toFixed()}K`
+  }
+  return `${value.toFixed()}`
+}
+
+export const parseInterval = (max: number, min: number) => {
+  const count = isMobile() ? 5 : 20
+  const interval = (max - min) / count
+  const { length } = Math.ceil(interval).toString()
+  const factor = 10 ** (length > 2 ? length - 2 : 0)
+  return (Math.ceil(interval / factor) + 1) * factor
+}
